uploader/requirements.txt
sanic==22.6.1
asyncpg==0.26.0
sqlalchemy==1.4.39
pandas==1.3.5
aiofiles==0.8.0
python-dotenv==0.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
